Fix for circleci compilation.

+++ src/mcts/stoppers/smooth.cc
@@ -326,7 +326,8 @@ class SmoothTimeManager : public TimeManager {
       move_allocated_time_ms_ = *time * params_.max_single_move_time_fraction();
     }
     if (move_allocated_time_ms_ > *time - params_.move_overhead_ms()) {
-      move_allocated_time_ms_ = std::max(0LL, *time - params_.move_overhead_ms());
+      move_allocated_time_ms_ = std::max(static_cast<int64_t>(0LL),
+                                         *time - params_.move_overhead_ms());
     }
     piggybank_time_ += time_to_piggybank_ms;
